feat: Implement Authentication HOC and Apply to Index Page
This commit introduces a new Higher-Order Component (HOC) named 'withAuth', which is designed to handle user authentication more efficiently across the application. The primary purpose of this HOC is to check the authentication status of the user and redirect unauthenticated users to the sign-in page.

Key Changes:
1. Created 'withAuth' HOC: This component wraps around other components and leverages the useSession hook to check for the presence of a valid user session. If a session does not exist, the HOC redirects the user to the sign-in page.

2. Applied 'withAuth' to the Index Page: The index page is now wrapped with the 'withAuth' HOC. This integration ensures that only authenticated users can access the index page, enhancing the security and user experience of the application.

3. Enhanced Session Management: Improved the SessionProvider to handle session states more effectively, ensuring a smoother authentication flow and user experience.

This implementation is a step towards robust authentication management in the application, aligning with our goals of enhancing security and user experience